+++ examples/snippets/app/concepts/adapters/flags.tsx
@@ -6,5 +6,5 @@ const edgeConfigAdapter = createEdgeConfigAdapter(process.env.EDGE_CONFIG!);
 export const customAdapterFlag = flag<boolean>({
   key: 'custom-adapter-flag',
   description: 'Shows how to use a custom flags adapter',
-  adapter: edgeConfigAdapter(),
+  adapter: edgeConfigAdapter,
 });